Separate `lib` code from `bin` in `mining-device`
Isolating the library code into a `lib/mod.rs` and consume it in
`main.rs`.
